Fixed bps.to_bytes()
Still no application.

The action lengths all add up to equal target output. So the process of normalizing is clearly correct to the extent that the lengths are valid.

If the lengths are valid, the action operations are too.

Which also means that the data used for obtaining relative offsets is too.

I'll add more QOL in the action class for debugging purposes that I expect would have been a later update.

The only error I can see currently, is that we may be misinterpreting the relative offset information.

I can't understand yet how or why as I don't think the relative lengths are modified after calling the decode/encode process considering a relative offset of 0 is, while unlikely, quite acceptable functionally.

And the LSB is signed. Decoding and Encoding works so byte order is not a problem.

I think its quite safe to say we can normalize and encode a BPS file/object.

Sweet lord I can only imagine the graceless agony BPS construction will wield upon my heretical ambitious excretion hole.