Added to untested addition code
The lengths of the operations before and after still have not been added.

Had this been IPS this would not be a problem. Byuu really did a number on this format.

The range function is copied from patchlib.ips but it should work still.

Should these functions crop up alot I may add them to a "common" subpackage which may be accessed by other modules.