Further Developing bps build
Removed delete/create due to non-functioning designs.

This new bps build method (while not yet functioning) is much faster than the last design.

We are currently getting patch sizes about 80KB over what we are hoping to get.

Theory tells me that I can optimize the relative offsets further but I am not so sure that that is the answer.

Naturally bps is a logistical mess.